+++ Client.Core/Flux/Internal/IFluxResultMapper.cs
@@ -1,3 +1,4 @@
+using System;
 using InfluxDB.Client.Core.Flux.Domain;
 
 namespace InfluxDB.Client.Core.Flux.Internal
@@ -14,5 +15,13 @@ namespace InfluxDB.Client.Core.Flux.Internal
         /// <typeparam name="T">Type of DomainObject</typeparam>
         /// <returns>Converted DomainObject</returns>
         T ConvertToEntity<T>(FluxRecord fluxRecord);
+
+        /// <summary>
+        /// Converts FluxRecord to DomainObject specified by Type.
+        /// </summary>
+        /// <param name="fluxRecord">Flux record</param>
+        /// <param name="type">Type of DomainObject</param>
+        /// <returns>Converted DomainObject</returns>
+        object ConvertToEntity(FluxRecord fluxRecord, Type type);
     }
 }